Fixed a bug with StructDlg when editing settings
So if you are selecting multiple files in structdlg, then you decide you want to hit Arume->Edit Settings, then structdlg errors out, because it does not expect a cell array as a default value. Here, i catch this condition by checking if the default value is a cell array, then I convert to a string (so the user can view the file names in the dlg), but save the original cell array in the structdlg output (so the format is not changed when it is saved).